Gui: settle BacklightIntensity on one parameter type
A parameter group is a set of typed maps, so an Int and a Float of the
same name are two separate values. BacklightIntensity was read as both:
this class took a Float fraction, while the Clipping dialog's slider, the
3D view preference page and the viewer all took an Int per cent -- and
only the Int ever lit anything.

So ViewParams::getBacklightIntensity() returned a value nothing writes,
and since the Clipping dialog seeds its slider from it, that slider has
been coming up at 1 of 100 rather than at the stored intensity. The
accessor becomes the Int per cent everything else already agrees on, and
ViewParams::migrate() drops the Float slot, carrying a value over as a
per cent first where a config has one -- only scripting could have
written it. Removing it is what stops the migration repeating.

BacklightColor is Unsigned on both sides and needs nothing.

Verified against a config seeded with the old Float: 0.5 comes up as the
Int 50 and no Float of that name survives the start. On a current config,
moving the panel's slider to 40 writes the light factor 0.4 and the
preference 40, and the backlight follows.

+++ src/Gui/Clipping.cpp
@@ -384,7 +384,7 @@ Clipping::Clipping(Gui::View3DInventor* view, QWidget* parent)
     d->ui.checkBoxGroupRendering->initAutoSave(ViewParams::getSectionFillGroup());
     d->ui.checkBoxBacklight->initAutoSave(ViewParams::getEnableBacklight());
     d->ui.backlightColor->initAutoSave(static_cast<uint>(ViewParams::getBacklightColor()));
-    d->ui.sliderIntensity->initAutoSave(ViewParams::getBacklightIntensity());
+    d->ui.sliderIntensity->initAutoSave(static_cast<int>(ViewParams::getBacklightIntensity()));
     d->ui.checkBoxShowPlane->initAutoSave(ViewParams::getShowClipPlane());
     d->ui.spinBoxPlaneSize->initAutoSave(ViewParams::getClipPlaneSize());
     d->ui.editHatchTexture->setFileName(
